Adds different flavors of event (#1)
Gives a facelift to event, adding combinators for memoization.

In future versions of the library, we'll try to improve these combinators to make memoization even more efficient.
We've also dropped `Apply` and `Applicative` from `Event` because `Apply` and `Applicative` have their own notion of time that is not that of `Event`. From the [`Applicative` wiki](https://en.wikipedia.org/wiki/Applicative_functor):

> Applicative functors allow for functorial computations to be sequenced (unlike plain functors), but don't allow using results from prior computations in the definition of subsequent ones (unlike monads).

Events have their own unique idea of time, so adding type-classes that imply temporality (sequencing) is a mixed metaphor that leads to odd results. Specifically, in the legacy event, you could have `(map (add 1) x) <*> x` emit `3` when 1 is pushed to the event followed by 2 because it sequences the applicative from left to right. This leads to the curious result where `2+2` equals `3` before it eventually equals `4`, which is tough to justify in any conceptual framework.

If `Apply` were necessary to achieve these sorts of results then it'd be worth holding onto, but we already have `sampleOn`, which is the same as apply but makes it explicit which side's temporality we are choosing. In the examples  & in anger, we haven't found a single use of `Apply` that couldn't be achieved using `sampleOn`, and in many cases, using `sampleOn` eliminated some of these `2+2=3` results.

Getting rid of apply also gets rid of `Applicative`, and we would like to be able to send a result immediately upon subscription. Again, though, the choice of `pure` to send a result once as opposed to, say, send the result once a second, seems arbitrary. To resolve this, we added `bang`, which is defined as an event that is emitted once. This feels like a more event-friendly version of `pure`, as it is linked to `Event`'s temporality instead of `Applicative`'s.