Changes Behavior to Poll and makes events hot (#38)
Events are now hot hot, meaning that an event cannot run arbitrary side effects upon subscription.

Additionally, changes `Behavior` to `Poll`, adding new instances and changing the behavior of some existing instances. Check out [this article](https://dev.to/mikesol/a-poll-hot-events-or-applicatives-choose-two-4454) for more info.

`Poll`s and `Event`s are almost interchangeable - a `Poll` with a single incoming event is isomorphic to a hot event + an `Applicative` instance.